+++ src/ViewInstance.js
@@ -748,7 +748,7 @@ class ViewInstance extends React.Component {
 
     const canCentralTenantCreateOrder = checkIfUserInCentralTenant(stripes) && checkIfCentralOrderingIsActive(centralOrdering);
     const canCreateOrder = (!checkIfUserInCentralTenant(stripes) && stripes.hasInterface('orders') && stripes.hasPerm('ui-inventory.instance.order.create')) || canCentralTenantCreateOrder;
-    const canReorder = stripes.hasPerm('ui-requests.reorderQueue');
+    const canReorder = stripes.hasPerm('ui-requests.reorderQueue.execute');
     const canExportMarc = stripes.hasPerm('ui-data-export.edit');
     const canAccessLinkedDataOptions = stripes.hasPerm(LINKED_DATA_EDITOR_PERM);
     const isSourceLinkedData = isLinkedDataSource(source);