+++ src/settings/SubjectTypesSettings.js
@@ -26,7 +26,7 @@ const SubjectTypesSettings = (props) => {
   const stripes = useStripes();
 
   const hasPerm = stripes.hasPerm('ui-inventory.settings.subject-types');
-  const suppress = getSourceSuppressor([RECORD_SOURCE.FOLIO]);
+  const suppress = getSourceSuppressor([RECORD_SOURCE.FOLIO, RECORD_SOURCE.CONSORTIUM]);
 
   return (
     <TitleManager