+++ CHANGELOG.md
@@ -118,6 +118,7 @@ and disable fields when "Settings (Inventory): Create, edit and delete HRID hand
 * Disallow displaying shared instances when find instance modal is open. Refs UIIN-3072.
 * ECS: Item affiliation cannot be changed when it is attached to a local order. Refs UIIN-3063.
 * Refactor ui-inventory permissions. Refs UIIN-3087.
+* Disallow displaying shared instances when find instance modal is open. Refs UIIN-3072.
 
 ## [11.0.5](https://github.com/folio-org/ui-inventory/tree/v11.0.5) (2024-08-29)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v11.0.4...v11.0.5)