+++ CHANGELOG.md
@@ -122,6 +122,7 @@ and disable fields when "Settings (Inventory): Create, edit and delete HRID hand
 * Update permission name after Review and cleanup Module Descriptor for ui-requests. Refs UIIN-3058.
 * Browse | Number of titles in Subject browse results does not match the number of instances returned in search. Fixes UIIN-3101.
 * Adjust the URI of a redirect to Linked data editor. Fixes UIIN-3107.
+* Add 'replaces' array to refactored ui permissions. Refs UIIN-3110.
 
 ## [11.0.5](https://github.com/folio-org/ui-inventory/tree/v11.0.5) (2024-08-29)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v11.0.4...v11.0.5)