+++ CHANGELOG.md
@@ -125,6 +125,7 @@ and disable fields when "Settings (Inventory): Create, edit and delete HRID hand
 * Rename "mod-settings.global.*" permissions. Refs UIIN-3109.
 * Suppress edit and delete actions of subject types and sources for `consortium` source. Refs UIIN-3112.
 * Add 'replaces' array to refactored ui permissions. Refs UIIN-3110.
+* Update quick-marc permission name. Refs UIIN-3111.
 
 ## [11.0.5](https://github.com/folio-org/ui-inventory/tree/v11.0.5) (2024-08-29)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v11.0.4...v11.0.5)